requirements.txt
fastapi
google-generativeai
pydantic
python-dotenv
uvicorn[standard]